requirements.txt
azure-identity
Flask==2.3.2
openai==1.40.6
python-dotenv==1.0.1
azure-core
azure-search-documents==11.6.0b4
azure-ai-formrecognizer==3.3.3
Markdown==3.4.4
requests==2.31.0
tqdm==4.65.0
tiktoken==0.7.0
langchain==0.0.292
bs4==0.0.1
urllib3==2.0.6
pytest==7.4.0
pillow==10.4.0
azure-ai-documentintelligence==1.0.0b3
azure-ai-vision-imageanalysis==1.0.0b3
sentence-transformers==3.0.1